counter with redux working and organized
