+++ src/components/Counter/Counter.js
@@ -1,0 +1,17 @@
+import { connect } from 'react-redux';
+import { Container } from './Container';
+
+const mapStateToProps = state => {
+  return {
+    count: state
+  };
+};
+
+const mapDispatchToProps = dispatch => {
+  return {
+    handleIncrementClick: () => dispatch( { type: 'INCREMENT' } ),
+    handleDecrementClick: () => dispatch( { type: 'DECREMENT' } )
+  }
+};
+
+export const Counter = connect( mapStateToProps, mapDispatchToProps )( Container );
