+++ src/store.js
@@ -1,0 +1,4 @@
+import { createStore } from 'redux';
+import { countReducer } from './reducers/counter';
+
+export const store = createStore(countReducer);
